Remove CMS and switch from NPM to hugo.
This commit removes the CMS system completely.
While a nice idea, it added much complexity, which made the website
difficult to maintain and inconsistent when changes to the layout are made.
Additionally, the CMS and theme are no longer maintained by netlify.
This means we have to either use old hugo versions or painstakenly update
many dependent npm libraries ourselves.

Without the CMS we can also switch from using npm as the main
build plattform to hugo.
This makes the build process much faster and less complex.
All functionality (except the CMS) are retained.

NPM is still used to install postcss and its plugins.